+++ geotrek/maintenance/managers.py
@@ -1,13 +1,21 @@
-from django.db.models import Min, Max
-from django.db.models.functions import ExtractYear
+from django.db.models import Min, Max, Func, Case, When, IntegerField
+from django.db.models.functions import ExtractYear, Cast
 
 from geotrek.common.mixins.managers import NoDeleteManager
 
 
 class InterventionManager(NoDeleteManager):
     def year_choices(self):
-        values = self.existing().filter(begin_date__isnull=False).annotate(year=ExtractYear('begin_date')) \
-            .order_by('-year').distinct().values_list('year', flat=True)
+        """ Get all range years between begin_date and end_date and concatenates distinct years """
+        qs = (self.existing().all().annotate(
+            years=Func(
+                Cast(ExtractYear("begin_date"), output_field=IntegerField()),
+                Cast(Case(When(end_date__isnull=False, then=ExtractYear("end_date")),
+                          default=ExtractYear("begin_date")),
+                     output_field=IntegerField()),
+                function="generate_series"),
+        ))
+        values = qs.distinct('years').order_by('-years').values_list('years', flat=True)
         return [(year, year) for year in values]