+++ geotrek/maintenance/static/maintenance/main.js
@@ -32,3 +32,30 @@ $(window).on('entity:map', function (e, data) {
         }
     });
 });
+
+
+// Date picker with placeholder on input
+function setDatePickerConfig(idList) {
+    $(idList).datepicker({
+        autoclose: true,
+        language: window.SETTINGS.languages.default,
+        format: window.SETTINGS.date_format
+    });
+}
+$(window).on('entity:view:add', function (e, data) {
+    if (data.modelname === "intervention"){
+        setDatePickerConfig('#id_begin_date, #id_end_date');
+    };
+});
+
+$(window).on('entity:view:list', function (e, data) {
+    if (data.modelname === "intervention"){
+        setDatePickerConfig('#id_begin_date_0, #id_begin_date_1, #id_end_date_0, #id_end_date_1');
+    };
+});
+
+$(window).on('entity:view:update', function (e, data) {
+    if (data.modelname === "intervention"){
+        setDatePickerConfig('#id_begin_date, #id_end_date');
+    };
+});